+++ core/src/main/java/io/teamchallenge/core/OnlineStoreApplication.java
@@ -1,0 +1,13 @@
+package io.teamchallenge.core;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class OnlineStoreApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(OnlineStoreApplication.class, args);
+    }
+
+}
