+++ core/src/main/java/io/teamchallenge/core/controllers/HelloController.java
@@ -1,0 +1,15 @@
+package io.teamchallenge.core.controllers;
+
+import java.util.Optional;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class HelloController {
+
+    @GetMapping("/hello")
+    public String greetings(@RequestParam Optional<String> name){
+        return "Hello, " + name.orElse("Unknown");
+    }
+}
