+++ core/src/test/java/io/teamchallenge/core/CoreApplicationTests.java
@@ -1,0 +1,13 @@
+package io.teamchallenge.core;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CoreApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
