+++ dao/src/test/java/io/teamchallenge/dao/DaoApplicationTests.java
@@ -1,0 +1,13 @@
+package io.teamchallenge.dao;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class DaoApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
