+++ service-api/pom.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <parent>
+        <groupId>io.teamchallenge</groupId>
+        <artifactId>onlinestore</artifactId>
+        <version>0.0.1-SNAPSHOT</version>
+    </parent>
+    <groupId>io.teamchallenge</groupId>
+    <artifactId>service-api</artifactId>
+    <version>0.0.1-SNAPSHOT</version>
+    <name>service-api</name>
+    <description>service-api</description>
+</project>
