+++ service-api/src/test/java/io/teamchallenge/serviceapi/ServiceApiApplicationTests.java
@@ -1,0 +1,13 @@
+package io.teamchallenge.serviceapi;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ServiceApiApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
