--- src/main/java/io/teamchallenge/onlinestore/OnlineStoreApplication.java
@@ -1,13 +1,0 @@
-package io.teamchallenge.onlinestore;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class OnlineStoreApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(OnlineStoreApplication.class, args);
-    }
-
-}
