--- src/test/java/io/teamchallenge/onlinestore/OnlineStoreApplicationTests.java
@@ -1,13 +1,0 @@
-package io.teamchallenge.onlinestore;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class OnlineStoreApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
